move common folder into src direction
